Startinf moving settingsValidation to the settings area (#1895)
